Move functions which get assembled graphql schema to server-only code
