+++ packages/lesswrong/server/vulcan-lib/apollo-server/index.ts
@@ -1,3 +1,1 @@
 export * from './apollo_server';
-
-export { default as initGraphQL } from './initGraphQL';
